ISSUE-36 - Organize Environment and Logger init
- Moving config to its own file
