+++ cmd/discord_bot/main.go
@@ -4,6 +4,7 @@ import (
 	"log"
 
 	"github.com/caarlos0/env/v11"
+	config "github.com/ccil-kbw/robot/internal/config"
 	environment "github.com/ccil-kbw/robot/internal/environment"
 	logger "github.com/ccil-kbw/robot/internal/logger"
 	"github.com/ccil-kbw/robot/pkg/discord"
@@ -16,12 +17,12 @@ func init() {
 }
 
 func main() {
-	bot := discord.NewDiscordBot(logger.Logger, logger.Cfg.DiscordServerID, logger.Cfg.DiscordBotToken, logger.Cfg.DiscordBotAsPublic)
+	bot := discord.NewDiscordBot(logger.Logger, config.Cfg.DiscordServerID, config.Cfg.DiscordBotToken, config.Cfg.DiscordBotAsPublic)
 	bot.StartBot()
 }
 
 func loadConfig() {
-	if err := env.Parse(&logger.Cfg); err != nil {
+	if err := env.Parse(&config.Cfg); err != nil {
 		log.Fatalf("%+v\n", err)
 	}
 	log.Println("configuration ready")
